Extend validation of element attributes
