--- src/core/consts.js
@@ -1,326 +1,0 @@
-/*
-Copyright 2017-2018 Opera Software AS
-
-Licensed under the Apache License, Version 2.0 (the "License");
-you may not use this file except in compliance with the License.
-You may obtain a copy of the License at
-
-     http://www.apache.org/licenses/LICENSE-2.0
-
-Unless required by applicable law or agreed to in writing, software
-distributed under the License is distributed on an "AS IS" BASIS,
-WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-See the License for the specific language governing permissions and
-limitations under the License.
-*/
-
-{
-  const SUPPORTED_EVENTS = [
-    // mouse events
-    'onAuxClick',
-    'onClick',
-    'onContextMenu',
-    'onDoubleClick',
-    'onDrag',
-    'onDragEnd',
-    'onDragEnter',
-    'onDragExit',
-    'onDragLeave',
-    'onDragOver',
-    'onDragStart',
-    'onDrop',
-    'onMouseDown',
-    'onMouseEnter',
-    'onMouseLeave',
-    'onMouseMove',
-    'onMouseOut',
-    'onMouseOver',
-    'onMouseUp',
-    // keyboard events
-    'onKeyDown',
-    'onKeyPress',
-    'onKeyUp',
-    // focus events
-    'onFocus',
-    'onBlur',
-    // form events
-    'onChange',
-    'onInput',
-    'onInvalid',
-    'onSubmit',
-    // clipboard events
-    'onCopy',
-    'onCut',
-    'onPaste',
-    // composition events
-    'onCompositionEnd',
-    'onCompositionStart',
-    'onCompositionUpdate',
-    // selection events
-    'onSelect',
-    // touch events
-    'onTouchCancel',
-    'onTouchEnd',
-    'onTouchMove',
-    'onTouchStart',
-    // UI events
-    'onScroll',
-    // wheel events
-    'onWheel',
-    // media events
-    'onAbort',
-    'onCanPlay',
-    'onCanPlayThrough',
-    'onDurationChange',
-    'onEmptied',
-    'onEncrypted',
-    'onEnded',
-    'onError',
-    'onLoadedData',
-    'onLoadedMetadata',
-    'onLoadStart',
-    'onPause',
-    'onPlay',
-    'onPlaying',
-    'onProgress',
-    'onRateChange',
-    'onSeeked',
-    'onSeeking',
-    'onStalled',
-    'onSuspend',
-    'onTimeUpdate',
-    'onVolumeChange',
-    'onWaiting',
-    // image events
-    'onLoad',
-    'onError',
-    // animation events
-    'onAnimationStart',
-    'onAnimationEnd',
-    'onAnimationIteration',
-    // transition events
-    'onTransitionEnd',
-    // search event
-    'onSearch',
-    // toogle event
-    'onToggle',
-  ];
-
-  const SUPPORTED_ATTRIBUTES = [
-    // most used attributes
-    'tabIndex',
-    'href',
-    'draggable',
-    'name',
-    'disabled',
-    'type',
-    'value',
-    'id',
-    'checked',
-    'contentEditable',
-    'readOnly',
-    'alt',
-    'title',
-    'width',
-    'height',
-    'required',
-    'for',
-    'label',
-    'minLength',
-    'maxLength',
-    'method',
-    'src',
-    'rel',
-
-    // other attributes
-    'accept',
-    'acceptCharset',
-    'accessKey',
-    'action',
-    'allowFullScreen',
-    'allowTransparency',
-    'async',
-    'autoComplete',
-    'autoFocus',
-    'autoPlay',
-    'capture',
-    'cellPadding',
-    'cellSpacing',
-    'challenge',
-    'charSet',
-    'cite',
-    'classID',
-    'colSpan',
-    'cols',
-    'content',
-    'contextMenu',
-    'controls',
-    'coords',
-    'crossOrigin',
-    'data',
-    'dateTime',
-    'default',
-    'defer',
-    'dir',
-    'download',
-    'encType',
-    'form',
-    'frameBorder',
-    'headers',
-    'hidden',
-    'high',
-    'hrefLang',
-    'httpEquiv',
-    'icon',
-    'incremental',
-    'inputMode',
-    'integrity',
-    'is',
-    'keyParams',
-    'keyType',
-    'kind',
-    'lang',
-    'list',
-    'loop',
-    'low',
-    'manifest',
-    'marginHeight',
-    'marginWidth',
-    'max',
-    'media',
-    'mediaGroup',
-    'min',
-    'multiple',
-    'muted',
-    'noValidate',
-    'nonce',
-    'open',
-    'optimum',
-    'pattern',
-    'placeholder',
-    'poster',
-    'preload',
-    'profile',
-    'radioGroup',
-    'reversed',
-    'role',
-    'rowSpan',
-    'rows',
-    'sandbox',
-    'scope',
-    'scoped',
-    'scrolling',
-    'seamless',
-    'selected',
-    'shape',
-    'size',
-    'sizes',
-    'span',
-    'spellCheck',
-    'srcDoc',
-    'srcLang',
-    'srcSet',
-    'start',
-    'step',
-    'summary',
-    'target',
-    'useMap',
-    'wmode',
-    'wrap',
-
-    // aria attributes
-    'ariaActiveDescendant',
-    'ariaAtomic',
-    'ariaAutoComplete',
-    'ariaBusy',
-    'ariaChecked',
-    'ariaControls',
-    'ariaDescribedBy',
-    'ariaDisabled',
-    'ariaDropEffect',
-    'ariaExpanded',
-    'ariaFlowTo',
-    'ariaGrabbed',
-    'ariaHasPopup',
-    'ariaHidden',
-    'ariaInvalid',
-    'ariaLabel',
-    'ariaLabelLedBy',
-    'ariaLevel',
-    'ariaLive',
-    'ariaMultiLine',
-    'ariaMultiSelectable',
-    'ariaOrientation',
-    'ariaOwns',
-    'ariaPosInSet',
-    'ariaPressed',
-    'ariaReadOnly',
-    'ariaRelevant',
-    'ariaRequired',
-    'ariaSelected',
-    'ariaSetSize',
-    'ariaSort',
-    'ariaValueMax',
-    'ariaValueMin',
-    'ariaValueNow',
-    'ariaValueText',
-  ];
-
-  const getSupportedStyles = element => {
-    const keys = Object.keys(element.style);
-    if (keys.length) {
-      return keys;
-    }
-    return Object.keys(Object.getPrototypeOf(element.style))
-        .filter(key => !key.includes('-'));
-  };
-
-  const SUPPORTED_STYLES = getSupportedStyles(document.documentElement);
-
-  const SUPPORTED_FILTERS = [
-    'blur',
-    'brightness',
-    'contrast',
-    'dropShadow',
-    'grayscale',
-    'hueRotate',
-    'invert',
-    'opacity',
-    'sepia',
-    'saturate',
-  ];
-
-  const SUPPORTED_TRANSFORMS = [
-    'matrix',
-    'matrix3d',
-    'translate',
-    'translate3d',
-    'translateX',
-    'translateY',
-    'translateZ',
-    'scale',
-    'scale3d',
-    'scaleX',
-    'scaleY',
-    'scaleZ',
-    'rotate',
-    'rotate3d',
-    'rotateX',
-    'rotateY',
-    'rotateZ',
-    'skew',
-    'skewX',
-    'skewY',
-    'perspective',
-  ];
-
-  const Consts = {
-    SUPPORTED_ATTRIBUTES,
-    SUPPORTED_EVENTS,
-    SUPPORTED_STYLES,
-    SUPPORTED_FILTERS,
-    SUPPORTED_TRANSFORMS,
-  };
-
-  module.exports = Consts;
-}
